refactor: combine pastoral council strapi files
