+++ astro-frontend/src/components/sections/landing-page/obituary-section.astro
@@ -92,6 +92,6 @@ const featuredObituary = obituaries.length === 5 ? obituaries[4] : null;
 			</div>
 		</>
 	) : (
-	<p class="py-8 text-center text-3xl font-semibold">{Message.OBITUARIES}</p>
+	<p class="py-8 text-center text-3xl font-semibold">{Message.OBITUARIES_NOT_FOUND}</p>
 	)}
 </section>
