+++ astro-frontend/src/components/sections/landing-page/parish-priest-speaks-section.astro
@@ -66,7 +66,7 @@ const parishPriestMessage =
         {
           parishPriestMessage === null ? (
             <p class="py-8 text-center text-3xl font-semibold">
-              {Message.MESSAGE}
+              {Message.MESSAGE_TO_BE_SHARED}
             </p>
           ) : (
             parishPriestMessage?.message
