+++ astro-frontend/src/components/sections/parish-pastoral-council/parish-pastoral-council-list.astro
@@ -1,5 +1,6 @@
 ---
 import NetworkImage from '~/components/ui/network-image.astro';
+import { getPlaceholderImage } from '~/helpers/get-placeholder-image';
 import type { ParishPastoralCouncil } from '~/models/parish-pastoral-council';
 
 const { pastoralCouncilData } = Astro.props as { pastoralCouncilData : ParishPastoralCouncil[] };
@@ -8,11 +9,14 @@ const { pastoralCouncilImage } = Astro.props;
 
 <div>
   <NetworkImage
-    src={pastoralCouncilImage?.pastoralCouncilImage.url ||
-      "Parish Pastoral Council of Our Lady of Pompei Church"}
-    alt="Parish Pastoral Council of Our Lady of Pompei Church"
-    width={pastoralCouncilImage?.pastoralCouncilImage.width || 300}
-    height={pastoralCouncilImage?.pastoralCouncilImage.height || 10}
+    src={pastoralCouncilImage.pastoralCouncilImage?.url
+      ? new URL(
+          pastoralCouncilImage.pastoralCouncilImage.url,
+          import.meta.env.STRAPI_URL
+        ).toString()
+      : getPlaceholderImage()}
+    width={pastoralCouncilImage.pastoralCouncilImage?.width || 600}
+    height={pastoralCouncilImage.pastoralCouncilImage?.height || 400}
     class="w-full rounded-sm object-contain pb-4 pt-4"
   />
   <table class="mx-auto w-full max-w-4xl border border-gray-200">
