+++ astro-frontend/src/constants/message.ts
@@ -1,7 +1,7 @@
 export class Message {
-  public static MESSAGE = 'To be shared soon.';
-  public static EVENTS = 'No Events Found!';
-  public static OBITUARIES = 'No Obituaries Found.';
-  public static COUNCIL = 'No Council Members Found.';
-  public static MAGAZINE = 'No Magazines available.';
+  public static MESSAGE_TO_BE_SHARED = 'Will be shared soon!';
+  public static EVENTS_NOT_FOUND = 'No Events Found!';
+  public static OBITUARIES_NOT_FOUND = 'No Obituaries Found!';
+  public static COUNCIL_MEMBERS_NOT_FOUND = 'No Council Members Found!';
+  public static MAGAZINES_NOT_FOUND = 'No Magazines available!';
 }
