--- astro-frontend/src/models/parish-pastoral-council-image.ts
@@ -1,57 +1,0 @@
-interface ImageFormat {
-  ext: string;
-  url: string;
-  hash: string;
-  mime: string;
-  name: string;
-  path: string | null;
-  size: number;
-  width: number;
-  height: number;
-  sizeInBytes: number;
-}
-
-interface ImageFormats {
-  large: ImageFormat;
-  small: ImageFormat;
-  medium: ImageFormat;
-  thumbnail: ImageFormat;
-}
-
-interface Media {
-  id: number;
-  documentId: string;
-  name: string;
-  alternativeText: string | null;
-  caption: string | null;
-  width: number;
-  height: number;
-  formats: ImageFormats;
-  hash: string;
-  ext: string;
-  mime: string;
-  size: number;
-  url: string;
-  previewUrl: string | null;
-  provider: string;
-  provider_metadata: null;
-  createdAt: string;
-  updatedAt: string;
-  publishedAt: string;
-}
-
-interface ParishPastoralCouncilImage {
-  id: number;
-  documentId: string;
-  createdAt: string;
-  updatedAt: string;
-  publishedAt: string;
-  pastoralCouncilImage: Media;
-}
-
-interface ParishPastoralCouncilImageData {
-  data: ParishPastoralCouncilImage;
-  meta: Record<string, unknown>;
-}
-
-export type { ParishPastoralCouncilImage, ParishPastoralCouncilImageData };
