+++ astro-frontend/src/models/parish-pastoral-council.ts
@@ -23,4 +23,58 @@ interface ParishPastoralCouncilData {
 	};
 }
 
-export type { ParishPastoralCouncilData };
+interface ImageFormat {
+  ext: string;
+  url: string;
+  hash: string;
+  mime: string;
+  name: string;
+  path: string | null;
+  size: number;
+  width: number;
+  height: number;
+  sizeInBytes: number;
+}
+
+interface Media {
+  id: number;
+  documentId: string;
+  name: string;
+  alternativeText: string | null;
+  caption: string | null;
+  width: number;
+  height: number;
+  formats: {
+		thumbnail: ImageFormat;
+		small: ImageFormat;
+		medium: ImageFormat;
+		large: ImageFormat;
+	};
+  hash: string;
+  ext: string;
+  mime: string;
+  size: number;
+  url: string;
+  previewUrl: string | null;
+  provider: string;
+  provider_metadata: null;
+  createdAt: string;
+  updatedAt: string;
+  publishedAt: string;
+}
+
+interface ParishPastoralCouncilImage {
+  id: number;
+  documentId: string;
+  createdAt: string;
+  updatedAt: string;
+  publishedAt: string;
+  pastoralCouncilImage: Media;
+}
+
+interface ParishPastoralCouncilImageData {
+  data: ParishPastoralCouncilImage;
+  meta: Record<string, unknown>;
+}
+
+export type { ParishPastoralCouncilData, ParishPastoralCouncilImage, ParishPastoralCouncilImageData };
