+++ astro-frontend/src/pages/events.astro
@@ -45,7 +45,7 @@ const pageTitle = `Page ${currentPage} | Events | Pompei Church`;
           </>
         ) : (
           <p class="py-8 text-center text-3xl font-semibold">
-            {Message.EVENTS}
+            {Message.EVENTS_NOT_FOUND}
           </p>
         )
       }
