+++ astro-frontend/src/pages/kok/events.astro
@@ -42,7 +42,7 @@ const pageTitle = `Page ${currentPage} | Events | Pompei Church`;
           </>
         ) : (
           <p class="py-8 text-center text-3xl font-semibold">
-            {Message.EVENTS}
+            {Message.EVENTS_NOT_FOUND}
           </p>
         )
       }
