+++ astro-frontend/src/pages/kok/obituary.astro
@@ -67,7 +67,7 @@ const obituaryUrl = new URL(
           </>
         ) : (
           <p class="py-8 text-center text-3xl font-semibold">
-            {Message.OBITUARIES}
+            {Message.OBITUARIES_NOT_FOUND}
           </p>
         )
       }
